Update README.md
What is FarSwap?

FarSwap is a software running on ethereum Blockhain that seeks to incentivize a network of users to operate a platform where users can Lend, borrow, buy and sell Crypto assets and earn massive rewards.

FarSwap is the first decentralized platform to give rewards to lenders and borrowers on the platform. 
FarSwap is an automated market maker (AMM) decentralized exchange (Dex) given by community vote. Users get 50% of the generated.

FarSwap Farming
Earn a minimum 55% APY 
unstake anytime you wish.
FarSwap Lending and borrowing
Lend and Borrow without KYC Get rewarded with Far token whenever you lend or borrow on FarSwap platform.
FarSwap is a protocol that reward participants extensively for their activities in an algorithm autonomous ecosystem without KYC. 
FarSwap distribute 50% trading fee to Stakers and liquidity providers on FarSwap platform.
FarSwap lending and borrowing
Get rewarded each time you lend or borrow on our platform. Pay back loan with FAR and get %2 repayment discount plus other rewards when you pay back loan with Far token.

Incentives for liquidity providers
Uniswap makews it possible for liquidity providers only earn transaction fees from the fund pool when they actively provide liquidity. Once they withdraw funds from the pool, they will no longer receive that passive income. In addition, as agreements become more and more popular, even though they are early liquidity providers, they still face being joined by large funds, exchanges, mining pools, and other (larger and richer) stakeholders to join the agreement, which will The risk of earnings dilution.
In FarSwap, liquidity is provided to the fund pool, and FAR tokens can also be rewarded. However, unlike Uniswap, even if you decide to no longer provide liquidity, the FAR tokens you get will also entitle you to continue to receive a portion of the fees in the FAR agreement. As an early miner involved in providing liquidity, you will become an important stakeholder in the agreement.
Income you receive from your investment in the mining in the LP ( provides flow mobility) the number of tokens proportional to the total number of coins on behalf of the LP. Unless you continue to increase the provision of liquidity, your holdings and corresponding reward income will gradually be diluted as the capital pool grows.

Incentives for Lenders and Borrowers
Far lending and borrowing platform  is developed and programmed for users to receive FAR token for reward anytime they lend or borrow cryptocurrency on the platform. users earn 2 %additional bonus when they choose to pay back loan with FAR token. users can stake the rewards they earn on FarSwap to earn more FAR token or exchange the FAR token to usdt.  users is requird to lock certain amount of FAR token befor they can borrow funds on the platform. No KYC is needed for this operation. Lend and borrow with little or no fee in return. users can also borrow or lend any crytocurrencies of their choice.

Token distribution

We know that many of us are existing liquidity providers in the Uniswap pool. Based on this situation, we have designed a token distribution mechanism to make it as easy as possible for existing Uniswap liquidity providers to migrate to our agreement!

In order to be able to start providing liquidity and obtain FAR tokens, anyone who holds Uniswap LP tokens can put these LP tokens in the corresponding initial pool list. Once this operation is completed, they will start earning token rewards after the block height of 287349823 . The list of LP tokens can be added by way of governance voting on the chain. Therefore, this will be a decision made by each of us .

In each block, 10 tokens will be created . These tokens will be evenly distribute
to the miners in each fund pool.

However, the number of FAR tokens generated in the first 200,000 blocks (about 2 weeks) will be 10 times the usual amount, that is, 100 SUSHI tokens will be rewarded in each block.
This is also to motivate the early farmers and participants of the agreement, and to provide assistance for liquidity migration.
The amount of FAR/WETH pool is twice that of others, so you are welcome to put your FAR  tokens into the fund pool of uniswap to get more “delicious”. After FAR is online, the community can vote to add more mining pools or change the FAR weight of any pool. How to change, you decide!
Reward distribution
According to the current Uniswap configuration, 0.3% of all transaction fees in any pool will be distributed proportionally to the liquidity providers in the pool. In FarSwap, 0.35% goes directly to active liquidity providers, and the remaining 0.08% will be exchanged for FAR(obviously through FarSwap) and distributed to FAR token holders.
Protocol usage
At the protocol and smart contract level, SushiSwap and FarSwap share the same interface. This means that if your protocol is currently compatible with Sushiswap, integration with FarSwap should be relatively simple. We will start integrating FarSwap with various DeFi and other ecosystem tools.

Security audit
Safety is very important to us. We have done unit testing as much as possible and tested everything that can be done on the testnet.
The contract you are about to review will contain all the content in the FarSwap smart contract repository, which is available from https://github.com/FarSwap    FarSwap-repo • GitHub and is currently available for review.
At the same time, MasterChef’s “agricultural” contract is unaudited, but for most Solidity developers, it is simple. Before you use LP tokens, please read the contract carefully! Everything is in the beta, please do so at your own risk.